refactor: Reorganizes files to enhance the analytics module.
