--- src/app/modules/knowledge-analytics/kc-statistics/events-table/events-table.component.html
@@ -1,36 +1,0 @@
-<div class="table-container">
-    <table mat-table [dataSource]="events" class="mat-elevation-z2">
-        <ng-container matColumnDef="timeStamp">
-            <th mat-header-cell *matHeaderCellDef> Timestamp </th>
-            <td mat-cell *matCellDef="let element"> {{element.timeStamp.toUTCString()}} </td>
-        </ng-container>
-
-        <ng-container matColumnDef="type">
-            <th mat-header-cell *matHeaderCellDef> Type </th>
-            <td mat-cell *matCellDef="let element"> {{element.type}} </td>
-        </ng-container>
-
-        <ng-container matColumnDef="kcId">
-            <th mat-header-cell *matHeaderCellDef> Knowledge Component Id </th>
-            <td mat-cell *matCellDef="let element"> {{element.knowledgeComponentId}} </td>
-        </ng-container>
-
-        <ng-container matColumnDef="learnerId">
-            <th mat-header-cell *matHeaderCellDef> Learner Id </th>
-            <td mat-cell *matCellDef="let element"> {{element.learnerId}} </td>
-        </ng-container>
-
-        <ng-container matColumnDef="specificData">
-            <th mat-header-cell *matHeaderCellDef width="50%"> Event-specific data </th>
-            <td mat-cell *matCellDef="let element" width="50%"> {{element.specificData | json}} </td>
-        </ng-container>
-
-        <tr mat-header-row *matHeaderRowDef="displayedColumns; sticky: true"></tr>
-        <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-    </table>
-</div>
-<button id="exportToCSV" color="primary" mat-raised-button (click)="exportToCSV()">Export to CSV</button>
-<button id="exportAllToCSV" color="accent" mat-raised-button (click)="exportAllToCSV()">Export all to CSV</button>
-<mat-paginator (page)="changePage($event)" [length]="count"
-    [pageSizeOptions]="[20, 50, 100]" [pageSize]="pageSize">
-</mat-paginator>
