--- src/app/modules/knowledge-analytics/kc-statistics/events-table/events-table.component.scss
@@ -1,16 +1,0 @@
-table {
-    width: 100%;
-}
-
-.table-container {
-  max-height: calc(100vh - 115px);
-  overflow: auto;
-}
-
-#exportToCSV {
-  margin: 10px;
-}
-
-#exportAllToCSV {
-  margin: 10px;
-}
