+++ src/app/modules/knowledge-analytics/kc-statistics/kc-statistics.component.html
@@ -1,5 +1,5 @@
-<div class="right-margin">
-    <mat-card fxLayout="row" fxLayoutAlign="center center" class="mat-elevation-z5 margin sticky">
+<div>
+    <mat-card fxLayout="row" fxLayoutAlign="center center" class="mat-elevation-z5 sticky">
         <span fxFlex>
             <mat-form-field appearance="fill">
                 <mat-label>Lekcije</mat-label>
