+++ src/app/modules/knowledge-analytics/kc-statistics/kc-statistics.component.scss
@@ -1,11 +1,3 @@
-.margin {
-  margin: 10px;
-}
-
-.right-margin {
-  margin-right: 70px;
-}
-
 .sticky {
   position: sticky;
   top: 0;
