+++ src/app/modules/knowledge-analytics/model/group.model.ts
@@ -1,7 +1,4 @@
-import {Learner} from './learner.model';
-
 export interface Group {
   id: number;
   name: string;
-  learners: Learner[];
 }
