+++ src/app/modules/monitoring/model/learner-progress.model.ts
@@ -1,4 +1,4 @@
-import {Learner} from '../../knowledge-analytics/model/learner.model';
+import {Learner} from './learner.model';
 import {KnowledgeComponentProgress} from './knowledge-component-progress.model';
 
 export interface LearnerProgress {
