+++ src/app/modules/monitoring/model/learner.model.ts
